:art: Adapt user schemas to user models. - Moved `TokenSchema`, `LoginSchema` and `RegisterSchema` to `user.py`. - Update user fields to new user model. - Update documentation. - Fixed imports.
