+++ src/common/models/__init__.py
@@ -1,9 +1,14 @@
-from src.common.models.auth import LoginSchema, RegisterSchema, TokenSchema
 from src.common.models.client import Client, ClientInsert
 from src.common.models.message import Message, MessageInsert, MessageType
 from src.common.models.responses import GenericResponse, create_response
 from src.common.models.ticket import Ticket
-from src.common.models.user import SystemRole, User
+from src.common.models.user import (
+    LoginSchema,
+    RegisterSchema,
+    SystemRole,
+    TokenSchema,
+    User,
+)
 from src.common.models.webhook import TwilioWebHook
 
 __all__ = [
