+++ src/common/models/user.py
@@ -1,8 +1,11 @@
 from datetime import datetime
 from enum import Enum
-from typing import Any
+from typing import Any, Optional
+from uuid import UUID
 
 from pydantic import BaseModel, EmailStr, Field, field_serializer
+
+from src.common.utils import PASSWORD_PATTERN
 
 
 class SystemRole(str, Enum):
@@ -12,11 +15,33 @@ class SystemRole(str, Enum):
     SUPPORT = "SUPPORT"
 
 
-class User(BaseModel):
-    id: int
+class TokenSchema(BaseModel):
+    access_token: str
+    refresh_token: str = "not implemented"
+
+
+class LoginSchema(BaseModel):
+    email: EmailStr
+    password: str = Field(min_length=8, max_length=32, pattern=PASSWORD_PATTERN)
+
+
+class RegisterSchema(BaseModel):
     email: EmailStr
     role: SystemRole = Field(default=SystemRole.WORKER)
+    password: str = Field(min_length=8, max_length=32, pattern=PASSWORD_PATTERN)
+    names: Optional[str] = Field(min_length=4, max_length=40, default=None,)
+    lastnames: Optional[str] = Field(min_length=4, max_length=40, default=None)
+
+
+class User(BaseModel):
+    id: UUID
+    email: EmailStr
     password: str
+    names: Optional[str]
+    lastnames: Optional[str]
+    system_role: SystemRole
+    job_role_id: Optional[UUID]
+    active: bool
     created_at: datetime
     updated_at: datetime
